removendo coisas desnecessarias e adaptando o projeto pro mongo

+++ backend/src/useCase/deleteEvent/deleteEventUseCase.ts
@@ -4,7 +4,11 @@ export class DeleteEventUseCase{
     constructor(private eventRepository:IEventRepository){}
 
     async execute(id:string){
-        let eventDeleted = await this.eventRepository.deleteEvent(id);
+        let eventExist = await this.eventRepository.findById(id);
+        if(!eventExist){
+            throw Error("Evento com o ID passado não existe!");
+        }
+        await this.eventRepository.deleteEvent(id);
         
         return "Evento deletado com sucesso!";
     }
